카페 마이페이지_ 작업 중[BCMPS]

+++ cafeback/AroundCafe/src/main/java/com/example/demo/mypage/cafe/repository/cafe/CafeRepository.java
@@ -11,5 +11,5 @@ public interface CafeRepository extends JpaRepository<Cafe, Long> {
     Optional<Cafe> findOne();
 
     @Query(value = "select * from cafe where member_no : memNo",nativeQuery = true)
-    Optional<Cafe> findByMemberNo(Integer memNo);
+    Optional<Cafe> findByMemberNo(Long memNo);
 }
